refactor: move course-highlight.directive to it own directory
