+++ src/app/pages/courses-page/pages/courses-list-page/directives/course-highlight/course-highlight.directive.spec.ts
